requirements.txt
connexion < 2.0.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
SQLAlchemy
SQLAlchemy-Utils
Flask-SQLAlchemy
sqlathanor
